Update all models to have long id instead of int | Make some adjustments for saving team as single transaction

+++ app/src/main/java/ua/hospes/rtm/db/cars/CarEntity.kt
@@ -7,7 +7,7 @@ import ua.hospes.rtm.domain.cars.Car
 
 @Entity(tableName = "cars")
 data class CarEntity(
-        @PrimaryKey(autoGenerate = true) val uid: Int = 0,
+        @PrimaryKey(autoGenerate = true) val uid: Long = 0,
         @ColumnInfo(name = "number") val number: Int,
         @ColumnInfo(name = "quality") val quality: Car.Quality = Car.Quality.NORMAL,
         @ColumnInfo(name = "is_broken") val broken: Boolean = false
